+++ contrib/criterion/__init__.py
@@ -1,6 +1,6 @@
 import torch
 from typing import List, Union
-from catalyst.utils import Factory, register
+from catalyst.contrib import Factory, register
 
 from . import unet as unet_loss
 from . import center_loss
